encapsulate
